[UPG]-Upgrade module l10n_cr_hr to Odoo v10
