--- l10n_cr_hr/README.rst
@@ -1,25 +1,0 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-   :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-   :alt: License: AGPL-3
-
-==========
-l10n_hr_cr
-==========
-
-This module add constraint to unique Nº Identification for employee
-
-
-Contributors
-------------
-
-* Lesmed Gutiérrez <lesmed.gutierrez@clearcorp.co.cr>
-
-
-Maintainer
-----------
-
-.. image:: https://avatars0.githubusercontent.com/u/7594691?v=3&s=200
-   :alt: ClearCorp
-   :target: http://clearcorp.cr
-
-This module is maintained by ClearCorp.
