+++ l10n_cr_hr/__openerp__.py
@@ -1,17 +1,40 @@
 # -*- coding: utf-8 -*-
-# © 2016 ClearCorp
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
+##############################################################################
+#
+#    OpenERP, Open Source Management Solution
+#    Addons modules by CLEARCORP S.A.
+#    Copyright (C) 2009-TODAY CLEARCORP S.A. (<http://clearcorp.co.cr>).
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
 {
     'name': 'l10n_cr_hr',
-    'summary': """
-        Constraint for employee identification
+    'description': """ 
+l10n_cr_hr
+=======================
+This module add constraint to unique Nº Identification for employee
     """,
-    'version': '9.0.1.0',
+    'version': '1.0',
     'author': 'ClearCorp',
     'category': 'Human Resources',
-    'website': "http://clearcorp.cr",
-    'depends': ['hr'],
+    'website': "http://clearcorp.co.cr",
+    'depends': ['hr'
+                ],
+    'data': [
+            
+                     ],
     'active': False,
     'installable': True,
     'license': 'AGPL-3',
