--- l10n_cr_hr/i18n/l10n_cr_hr.pot
@@ -1,33 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* l10n_cr_hr
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 9.0e\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2016-04-08 20:13+0000\n"
-"PO-Revision-Date: 2016-04-08 20:13+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: l10n_cr_hr
-#: code:addons/l10n_cr_hr/modules/hr_employee.py:17
-#, python-format
-msgid " (Copy)"
-msgstr " (Copia)"
-
-#. module: l10n_cr_hr
-#: model:ir.model,name:l10n_cr_hr.model_hr_employee
-msgid "Employee"
-msgstr "Empleado"
-
-#. module: l10n_cr_hr
-#: sql_constraint:hr.employee:0
-msgid "The Nº Identification already exist for other employee"
-msgstr "El Nº Identifiación ya existe para otro empleado"
-
