--- l10n_cr_hr/modules/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-# © 2016 ClearCorp
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-import hr_employee
